requirements.txt
Django==1.7
django-cors-headers==1.1.0
django-filter==0.10.0
django-rest-swagger==0.3.2
djangorestframework==3.1.2
envvars==0.3.0
Markdown==2.6.2
PyYAML==3.11
